refactor: extract delayed loading state logic into useDelayedLoadingState hook
