+++ src/lib/hooks.ts
@@ -1,4 +1,4 @@
-import { Inputs, useEffect, useReducer, useRef } from 'preact/hooks'
+import { Inputs, useEffect, useReducer, useRef, useState } from 'preact/hooks'
 
 export const useRedraw = () => {
   const [, redraw] = useReducer(c => c + 1, 0)
@@ -23,4 +23,28 @@ export const useUpdatingRef = <T>(value: T) => {
   const ref = useRef(value)
   ref.current = value
   return ref
+}
+
+export const useDelayedLoadingState = (loading: boolean, minLoadTime: number) => {
+  const [delayedLoading, setDelayedLoading] = useState(loading)
+  const loadingStartTime = useRef(0)
+
+  useEffect(() => {
+    if (loading) {
+      setDelayedLoading(true)
+      loadingStartTime.current = Date.now()
+    } else {
+      const elapsedTime = Date.now() - loadingStartTime.current
+      if (elapsedTime < minLoadTime) {
+        const remainingTime = Math.max(minLoadTime - elapsedTime, 0)
+        const id = setTimeout(() => {
+          setDelayedLoading(false)
+          loadingStartTime.current = 0
+        }, remainingTime)
+        return () => clearTimeout(id)
+      }
+    }
+  }, [loading, minLoadTime])
+
+  return delayedLoading
 }
